fix more bug and improve performance
